app/models/image_model.py
import cv2
import numpy as np
from flask_socketio import emit
from datetime import datetime

def analyze_image(file_path, session_id=None, progress_dict=None):
    """Analyze image for deepfake detection"""
    # Update progress
    if progress_dict is not None and session_id is not None:
        progress_dict[session_id] = 30
        emit('progress_update', {'progress': 30}, room=session_id)
    
    # Load image
    image = cv2.imread(file_path)
    
    # Update progress
    if progress_dict is not None and session_id is not None:
        progress_dict[session_id] = 50
        emit('progress_update', {'progress': 50}, room=session_id)
    
    # Simulate analysis (replace with actual model)
    # This is where you would integrate your deepfake detection model
    is_deepfake = np.random.random() > 0.7
    confidence = np.random.random() * 20 + (80 if is_deepfake else 75)
    
    # Update progress
    if progress_dict is not None and session_id is not None:
        progress_dict[session_id] = 80
        emit('progress_update', {'progress': 80}, room=session_id)
    
    # Generate anomalies if deepfake
    anomalies = []
    if is_deepfake:
        anomalies = [
            {
                "type": "Facial inconsistency",
                "severity": "high",
                "description": "Detected unnatural facial feature transitions"
            },
            {
                "type": "Texture anomaly",
                "severity": "medium",
                "description": "Inconsistent texture patterns in facial regions"
            }
        ]
    
    # Update progress
    if progress_dict is not None and session_id is not None:
        progress_dict[session_id] = 100
        emit('progress_update', {'progress': 100}, room=session_id)
    
    # Return result
    return {
        "isDeepfake": bool(is_deepfake),
        "confidence": float(confidence),
        "processingTime": float(np.random.random() * 2 + 1.5),
        "anomalies": anomalies,
        "mediaType": "image",
        "timestamp": datetime.now().isoformat()
    }